requirements.txt
hidapi
Pillow
pystray
